+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+### 2017-12-xx - Core update [v1.1.0]
+* GUI tool now lets you choose your Steam library folder (for those who set it in a separate place other than ~/.local/share/Steam/steamapps)
+* 
+
 ### 2017-12-11 - Database update
 * 2 games added to the database:
   - HITMAN™